Turn the ValuesCloseToTrendLineValidator back on for V and U sub-pattern scanning.
Also reversed the filtering for downtrends in both the VScanner and CupScanner. Previously filtering was being done on unique start and end date *before* the patterns were validated. This could have resulted in an invalid pattern being the filtered as the unique pattern, but then being disqualified for being invalid. This change will ensure only candidate sub-patterns which are valid are filtered for unique start and end dates.

Finally, in both VScanner and CupScanner, there was a duplicate initialization of the ValueCloseToTrendLineValidator. These duplicate initializations have been removed.